pyproject.toml
[project]
name = "sgb-advisor"
version = "0.1.0"
description = "Mathematically advises you on which SGB to buy"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.35.34",
    "pytest-playwright>=0.5.2",
    "python-dotenv>=1.0.1",
    "pyxirr>=0.10.5",
]

[tool.uv]
dev-dependencies = [
    "icecream>=2.1.3",
    "mypy>=1.11.2",
]
